Fix for links to local images (file:/// -> file://)
On Windows (I have not checked on other OSes), the path that MarkdownImages generates for local images is something like *file:///C:\...\image.png*.
This path is inserted in the html generated for the phantom to display the image.
But the result is a "broken link" image appearing in sublime (with the dimensions of the original image).
The culprit is the third / after 'file'. Removing it fixes the issue and makes the image load correctly.